@@ -13,7 +13,7 @@
     "private": true,
     "dependencies": {
         "@angular/animations": "^9.1.12",
-        "@angular/cdk": "^9.1.2",
+        "@angular/cdk": "^9.1.12",
         "@angular/common": "^9.1.12",
         "@angular/compiler": "^9.1.12",
         "@angular/core": "^9.1.12",
@@ -21,13 +21,8 @@
         "@angular/platform-browser": "^9.1.12",
         "@angular/platform-browser-dynamic": "^9.1.12",
         "@angular/router": "^9.1.12",
-        "@synerty/font-awesome": "^5.7.2",
-        "@synerty/ng2-balloon-msg": "^9.0.0",
-        "@synerty/ng2-balloon-msg-web": "^9.0.0",
-        "@synerty/peek-theme": "^0.2.5",
-        "@synerty/peek-util": "^9.0.0",
-        "@synerty/peek-util-web": "^9.0.0",
-        "@synerty/vortexjs": "^1.4.0",
+        "@synerty/peek-plugin-base-js": "^9.1.0",
+        "@synerty/vortexjs": "^2.0.2",
         "assert": "^1.5.0",
         "bootstrap": "^3.4.1",
         "core-js": "^3.6.5",
@@ -35,10 +30,8 @@
         "json-stable-stringify": "^1.0.1",
         "moment": "^2.24.0",
         "ng-zorro-antd": "^9.3.0",
-        "ng2-datetime-picker": "^0.13.2",
         "rxjs": "~6.5.3",
         "rxjs-compat": "^6.5.2",
-        "scss": "^0.2.4",
         "ts-md5": "^1.2.4",
         "zone.js": "~0.10.3"
     },
@@ -62,6 +55,6 @@
         "protractor": "^5.4.2",
         "ts-node": "~4.1.0",
         "tslint": "~5.9.1",
-        "typescript": "3.7.5"
+        "typescript": "3.8.3"
     }
 }
